Adding a package and making everything internal
