+++ Privalihin/Task03/Tests/binaryTrees/AVLTreeTest.kt
@@ -1,11 +1,12 @@
+package binaryTrees
+
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import kotlin.random.Random
 
-class BinarySearchTreeTest {
-
-    private fun randomBinarySearchTree(size: Int): BinarySearchTree<Int, Int> {
-        val tree = BinarySearchTree<Int, Int>()
+class AVLTreeTest {
+    private fun randomAVLTree(size: Int): AVLTree<Int, Int> {
+        val tree = AVLTree<Int, Int>()
 
         for (i in 1..size) {
             tree.insert(Random.Default.nextInt(), Random.Default.nextInt())
@@ -16,8 +17,7 @@ class BinarySearchTreeTest {
 
     @Test
     fun basicTest() {
-        val tree = BinarySearchTree<Int, Int>()
-
+        val tree = AVLTree<Int, Int>()
         for (i in 1..5) {
             tree.insert(i, i * i)
         }
@@ -32,14 +32,17 @@ class BinarySearchTreeTest {
     @Test
     fun correctnessTest() {
         for (i in 1..10) {
-            val tree = randomBinarySearchTree(Random.Default.nextInt(0, 10))
+            val tree = randomAVLTree(Random.Default.nextInt(0, 10))
+            assert(tree.isAVLTree())
             assert(tree.isBinarySearchTree())
         }
     }
 
     @Test
     fun stressTest() {
-        val tree = randomBinarySearchTree(1000)
+        val tree = randomAVLTree(1000)
+        assert(tree.parentsCorrectness())
+        assert(tree.isAVLTree())
         assert(tree.isBinarySearchTree())
     }
 }
