+++ Privalihin/Task03/Tests/binaryTrees/BinarySearchTreeTest.kt
@@ -1,10 +1,13 @@
+package binaryTrees
+
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import kotlin.random.Random
 
-class RBTreeTest {
-    private fun randomRBTree(size: Int): RBTree<Int, Int> {
-        val tree = RBTree<Int, Int>()
+class BinarySearchTreeTest {
+
+    private fun randomBinarySearchTree(size: Int): BinarySearchTree<Int, Int> {
+        val tree = BinarySearchTree<Int, Int>()
 
         for (i in 1..size) {
             tree.insert(Random.Default.nextInt(), Random.Default.nextInt())
@@ -15,7 +18,7 @@ class RBTreeTest {
 
     @Test
     fun basicTest() {
-        val tree = RBTree<Int, Int>()
+        val tree = BinarySearchTree<Int, Int>()
 
         for (i in 1..5) {
             tree.insert(i, i * i)
@@ -31,16 +34,14 @@ class RBTreeTest {
     @Test
     fun correctnessTest() {
         for (i in 1..10) {
-            val tree = randomRBTree(Random.Default.nextInt(0, 10))
-            assert(tree.isRBTree())
+            val tree = randomBinarySearchTree(Random.Default.nextInt(0, 10))
             assert(tree.isBinarySearchTree())
         }
     }
 
     @Test
     fun stressTest() {
-        val tree = randomRBTree(1000)
-        assert(tree.isRBTree())
+        val tree = randomBinarySearchTree(1000)
         assert(tree.isBinarySearchTree())
     }
 }
