+++ Privalihin/Task03/Tests/binaryTrees/RBTreeTest.kt
@@ -1,10 +1,12 @@
+package binaryTrees
+
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import kotlin.random.Random
 
-class AVLTreeTest {
-    private fun randomAVLTree(size: Int): AVLTree<Int, Int> {
-        val tree = AVLTree<Int, Int>()
+class RBTreeTest {
+    private fun randomRBTree(size: Int): RBTree<Int, Int> {
+        val tree = RBTree<Int, Int>()
 
         for (i in 1..size) {
             tree.insert(Random.Default.nextInt(), Random.Default.nextInt())
@@ -15,7 +17,8 @@ class AVLTreeTest {
 
     @Test
     fun basicTest() {
-        val tree = AVLTree<Int, Int>()
+        val tree = RBTree<Int, Int>()
+
         for (i in 1..5) {
             tree.insert(i, i * i)
         }
@@ -30,17 +33,16 @@ class AVLTreeTest {
     @Test
     fun correctnessTest() {
         for (i in 1..10) {
-            val tree = randomAVLTree(Random.Default.nextInt(0, 10))
-            assert(tree.isAVLTree())
+            val tree = randomRBTree(Random.Default.nextInt(0, 10))
+            assert(tree.isRBTree())
             assert(tree.isBinarySearchTree())
         }
     }
 
     @Test
     fun stressTest() {
-        val tree = randomAVLTree(1000)
-        assert(tree.parentsCorrectness())
-        assert(tree.isAVLTree())
+        val tree = randomRBTree(1000)
+        assert(tree.isRBTree())
         assert(tree.isBinarySearchTree())
     }
 }
