+++ Privalihin/Task03/src/binaryTrees/Color.kt
@@ -1,3 +1,5 @@
+package binaryTrees
+
 enum class Color {
     RED, BLACK
 }
